# src/main/java/io/swagger/api/NumbersApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.25).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.Numbers;
import io.swagger.model.SumResponse;
import io.swagger.model.Sums;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2021-04-08T21:52:52.495Z[GMT]")
@Validated
public interface NumbersApi {

    @Operation(summary = "Get the sum results", description = "", tags={ "Numbers" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Results successfully obtained", content = @Content(schema = @Schema(implementation = Sums.class))),
        
        @ApiResponse(responseCode = "404", description = "Results not found") })
    @RequestMapping(value = "/numbers",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Sums> getSums();


    @Operation(summary = "Insert a sum of two numbers.", description = "", tags={ "Numbers" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "201", description = "Numbers was successfully added", content = @Content(schema = @Schema(implementation = SumResponse.class))),
        
        @ApiResponse(responseCode = "400", description = "Bad Request - Check the request body"),
        
        @ApiResponse(responseCode = "500", description = "Internal Server Error") })
    @RequestMapping(value = "/numbers",
        produces = { "application/json" }, 
        consumes = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<SumResponse> insertSum(@Parameter(in = ParameterIn.DEFAULT, description = "The two numbers json object", required=true, schema=@Schema()) @Valid @RequestBody Numbers body);

}